python-service/tests/test_pdf.py
"""
Tests for PDF extraction service.
"""
import pytest
from fastapi import UploadFile
from io import BytesIO
from services.pdf_service import extract_text_from_pdf


@pytest.mark.asyncio
async def test_extract_text_from_valid_pdf():
    """Test PDF extraction with valid text-based PDF."""
    # This would need a real PDF file for integration testing
    # For now, this is a placeholder structure
    pass


@pytest.mark.asyncio
async def test_extract_text_from_empty_pdf():
    """Test PDF extraction with empty/scanned PDF."""
    # This would test handling of image-based PDFs
    pass


@pytest.mark.asyncio
async def test_extract_text_from_multipage_pdf():
    """Test PDF extraction with multiple pages."""
    pass


@pytest.mark.asyncio
async def test_invalid_file_type():
    """Test error handling for non-PDF files."""
    pass
