python-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
langchain==0.3.7
langchain-openai==0.2.8
openai==1.54.4
pdfplumber==0.11.4
sqlparse==0.5.1
python-multipart==0.0.17
aiofiles==24.1.0
